Move TestCaseValidator to separate package
